Rename onadata.celery to onadata.celeryapp
Hopefully stem circular import issue with prospector
