+++ onadata/__init__.py
@@ -11,6 +11,6 @@ __version__ = "3.3.1"
 
 # This will make sure the app is always imported when
 # Django starts so that shared_task will use this app.
-from .celery import app as celery_app
+from .celeryapp import app as celery_app
 
 __all__ = ("celery_app",)
