+++ onadata/apps/logger/import_tools.py
@@ -12,7 +12,7 @@ from django.core.files.uploadedfile import InMemoryUploadedFile
 from django.http.response import Http404
 
 from onadata.apps.logger.xform_fs import XFormInstanceFS
-from onadata.celery import app
+from onadata.celeryapp import app
 from onadata.libs.utils.logger_tools import create_instance
 
 # odk
