+++ onadata/celeryapp.py
